todo updated endpoint

+++ src/todo/todo.service.ts
@@ -3,6 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { Todo } from './interfaces/todo.interfaces';
 import { Model, Document } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
+import { User } from 'src/user/interfaces/user.interfaces';
 
 @Injectable()
 export class TodoService {
@@ -10,24 +11,32 @@ export class TodoService {
     @InjectModel('Todo') private readonly itemModel: Model<Todo & Document>,
   ) {}
 
-  async findAll(): Promise<Todo[]> {
-    return this.itemModel.find();
+  async findAll(user): Promise<Todo[]> {
+    return this.itemModel.find({ userId: user._id });
   }
 
-  async findOne(id: string): Promise<Todo> {
-    return this.itemModel.findOne({ _id: id });
+  async findOne(id: string, user): Promise<Todo> {
+    return this.itemModel.findOne({ _id: id, userId: user._id });
   }
 
-  async create(item: Todo): Promise<Todo> {
-    const newItem = new this.itemModel(item);
+  async create(todo: Todo, user): Promise<Todo> {
+    const userPayload = {
+      ...todo,
+      userId: user._id,
+    };
+    const newItem = new this.itemModel(userPayload);
     return await newItem.save();
   }
 
-  async update(id: string, item: Todo): Promise<Todo> {
-    return this.itemModel.findByIdAndUpdate(id, item, { new: true });
+  async update(id: string, item: Todo, user): Promise<Todo> {
+    return this.itemModel.findOneAndUpdate(
+      { _id: id, userId: user._id },
+      item,
+      { new: true },
+    );
   }
 
-  async delete(id: string): Promise<Todo> {
-    return this.itemModel.findByIdAndDelete(id);
+  async delete(id: string, user): Promise<Todo> {
+    return this.itemModel.findByIdAndDelete({ _id: id, userId: user._id });
   }
 }
